lazy git commit

+++ code/getHitPredict.R
@@ -4,7 +4,9 @@
 ## Download and compile HitPredict dababase.
 #------------------------------------------------------------------------------
 
-getHitPredict <- function(organism="HitPredict", downloads=getwd()) {
+getHitPredict <- function(organism="HitPredict", 
+			  downloads=getwd(), 
+			  keepdata=FALSE) {
 	suppressPackageStartupMessages({
 		require(dplyr)})
 	# Parse HitPredict Downloads page.
@@ -25,11 +27,13 @@ getHitPredict <- function(organism="HitPredict", downloads=getwd()) {
 		download.file(url,gzfile)
 		untar(gzfile, exdir=downloads)
 		rawdat <- data.table::fread(myfile,header=TRUE,skip=5)
-		unlink(myfile)
 	} else {
 		message("Using previously downloaded HitPredict data!")
 		untar(gzfile, exdir=downloads)
 		rawdat <- data.table::fread(myfile,header=TRUE,skip=5)
+	}
+	if (keepdata==FALSE) {
+		unlink(gzfile)
 		unlink(myfile)
 	}
 	# Clean up the raw data.
@@ -37,25 +41,7 @@ getHitPredict <- function(organism="HitPredict", downloads=getwd()) {
 	cleandat$Interactor_A_ID <- gsub("uniprotkb:","",cleandat$Interactor_A_ID)
 	cleandat$Interactor_B_ID <- gsub("uniprotkb:","",cleandat$Interactor_B_ID)
 	# Define organism specific mapping databases:
-	annotationDBs <- list(Anopheles_gambia = list(taxid=7165, database="org.Ag.eg.db", alias="anopheles"),
-			      Arabidopsis_thaliana = list(taxid=7302, database="org.At.tair.db", alias="arabidopsis"),
-			      Bos_taurus = list(taxid=9913, database="org.Bt.eg.db", alias="bovine"),
-			      Caenorhabditis_elegans = list(taxid=6239, database="org.Ce.eg.db",alias="worm"),
-			      Canis_familiaris = list(taxid=9615,database="org.Cf.eg.db",alias="dog"),
-			      Drosophila_melanogaster = list(taxid=7227,database="org.Dm.eg.db",alias="fly"),
-			      Danio_rerio = list(taxid=7955, database="org.Dr.eg.db",alias="zebrafish"),
-			      #Escherichia_coli = list(taxid=511145, database="org.EcK12.eg.db",alias="ecoli-k12"), # TaxID/uniprot mismatch? doesnt work
-			      Escherichia_coliSakai = list(taxid=386585, database="org.EcSakai.eg.db",alias="ecoli-sakai"),
-			      Gallus_gallus = list(taxid=9031, database="org.Gg.eg.db",alias="chicken"),
-			      Homo_sapiens = list(taxid=9606,database="org.Hs.eg.db",alias="human"),
-			      Mus_musculus = list(taxid=10090,database="org.Mm.eg.db",alias="mouse"),
-			      Macaca_mulatta = list(taxid=9544,database="org.Mmu.eg.db",alias="macaque"),
-			      Plasmodium_falciparum = list(taxid=5833,database="org.Pf.plasmo.db",alias="malaria"),
-			      Pan_troglodytes = list(taxid=9598,database="org.Pt.eg.db",alias="chimp"),
-			      Rattus_norvegicus = list(taxid=10116,database="org.Rn.eg.db",alias="rat"),
-			      Saccharomyces_cerevisiae = list(taxid=4932,database="org.Sc.sgd.db",alias="yeast"),
-			      Sus_scrofa = list(taxid=9823,database="org.Ss.eg.db",alias="pig"),
-			      Xenopus_laevis= list(taxid=8355,database="org.Xl.eg.db",alias="frog"))
+	source("annotationDBs.R")
 	# Subset HitPredict data, keep interactions from species with mapping databases...
 	dbs <- unlist(sapply(annotationDBs,"[",1))	
 	data <- data.table::as.data.table({
@@ -90,7 +76,7 @@ getHitPredict <- function(organism="HitPredict", downloads=getwd()) {
 		})
 		# Status report.
 		percent_mapped <- round(100*(1-sum(is.na(entrez))/length(entrez)),2)
-		message(paste("    ", percent_mapped, "% of", names(taxids)[taxids==species], 
+		message(paste("   ...", percent_mapped, "% of", names(taxids)[taxids==species], 
 			"Uniprot IDs were successfully mapped to Entrez IDs."))
 		# Create protein identifier map.
 		protMap <- as.list(entrez)
@@ -109,14 +95,18 @@ getHitPredict <- function(organism="HitPredict", downloads=getwd()) {
 						      "EntrezA","EntrezB"))
 	}
 	# Progress Report.
-	n <- dim(data)[1]
-	message(paste(n,"PPIs compiled from the HitPredict database and mapped to Entrez IDs."))
+	N <- dim(data)[1]
 	n_missing <- sum(is.na(data$EntrezA) | is.na(data$EntrezB))
-	percent_missing <- round(100*n_missing/n,2)
-	message(paste0(n_missing, " (", percent_missing, ") ", 
-		      "interactions are incompletely mapped to Entrez IDs and will be discarded."))
+	percent_missing <- round(100*n_missing/N,2)
+	n_missing <- format(as.numeric(n_missing), nsmall=0, big.mark=",")
+	message(paste0(n_missing, " (", percent_missing, "%) ", 
+		      "of all interactions (rows) are incompletely mapped to Entrez IDs and will be discarded."))
 	# Remove unmapped ids.
 	data <- dplyr::filter(data, !is.na(EntrezA) & !is.na(EntrezB))
+	n <- format(as.numeric(dim(data)[1]), nsmall=0, big.mark=",")
+	message(paste(n,"Protein-protein interactions from",
+		     length(taxids), "species were compiled from the HitPredict",
+		     "database and successfully mapped to Entrez IDs."))
 	# Return HitPredict data with Entrez IDs.
 	return(data)
 }
